Fixes for the second round of the code review
